feat(props): support most css props
Previously this was a shorthand-first library, but really a better
approach is to support most — if not all — commonly used CSS out of the box,
and provide only the most useful of them as shorthands.  Users can
easily augment this, and this change is not reductive.

Only notable change is that fe is now fw — macros take precendence over
shorthands, so they can overlap!